app[REFACT] moved initFetch in VM
