+++ packages/app/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./useFetchAuth";
 export * from "./useTextInput";
+export * as signInScreenVM from "./signInScreenVM";
